% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dimPlot}
\alias{dimPlot}
\title{visualize cells in 2D-dimensional space}
\usage{
dimPlot(
  object,
  colors.use = NULL,
  colors.ggplot = FALSE,
  reduction = NULL,
  group.by = NULL,
  split.by = NULL,
  shape.by = NULL,
  label = FALSE,
  shuffle = TRUE,
  seed = 1,
  pt.size = 1e-04,
  combine = TRUE,
  ...
)
}
\arguments{
\item{object}{seurat object}

\item{colors.use}{defining the color for each cell group}

\item{colors.ggplot}{whether use ggplot color scheme; default: colors.ggplot = FALSE}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{group.by}{Name of one or more metadata columns to group (color) cells by
(for example, orig.ident); pass 'ident' to group by identity class}

\item{split.by}{Name of a metadata column to split plot by;}

\item{shape.by}{If NULL, all points are circles (default).}

\item{label}{whether label the clusters in the 2D-space}

\item{shuffle}{Whether to randomly shuffle the order of points. This can be useful for crowded plots if points of interest are being buried.}

\item{seed}{Sets the seed if randomly shuffling the order of points.}

\item{pt.size}{size of the dots}

\item{combine}{Combine plots into a single patchworked ggplot object. If FALSE, return a list of ggplot objects}

\item{...}{Extra parameters passed to DimPlot from Seurat package}
}
\value{
ggplot2 object
}
\description{
visualize cells in 2D-dimensional space
}
