% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{doHeatmap}
\alias{doHeatmap}
\title{Feature expression heatmap}
\usage{
doHeatmap(
  object,
  features = NULL,
  cells = NULL,
  group.by = "ident",
  colors.use = NULL,
  colors.ggplot = FALSE,
  group.bar = TRUE,
  additional.group.by = NULL,
  additional.group.sort.by = NULL,
  color.heatmap = "RdBu",
  n.colors = 11,
  direction = -1,
  disp.min = -2.1,
  disp.max = 2.1,
  slot = "scale.data",
  assay = NULL,
  label = TRUE,
  text.size = 8,
  size = 4,
  hjust = 0,
  angle = 45,
  raster = TRUE,
  draw.lines = TRUE,
  lines.width = NULL,
  group.bar.height = 0.02,
  legend.remove = FALSE,
  legend.title = NULL,
  combine = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{A vector of features to plot, defaults to \code{VariableFeatures(object = object)}}

\item{cells}{A vector of cells to plot}

\item{group.by}{A vector of variables to group cells by; pass 'ident' to group by cell identity classes}

\item{colors.use}{A list of colors to use for the color bar. e.g., colors.use = list(ident = c("red","blue"))}

\item{colors.ggplot}{whether use ggplot color scheme; default: colors.ggplot = FALSE}

\item{group.bar}{Add a color bar showing group status for cells}

\item{additional.group.by}{A vector of variables to group cells by.}

\item{additional.group.sort.by}{which variable from additional.group.by to use for sorting cells}

\item{color.heatmap}{A character string or vector indicating the colormap option to use. It can be the avaibale color palette in viridis_pal() or brewer.pal()}

\item{n.colors}{number of basic colors to generate from color palette}

\item{direction}{Sets the order of colors in the scale. If 1, the default colors are used. If -1, the order of colors is reversed.}

\item{disp.min}{Minimum display value (all values below are clipped)}

\item{disp.max}{Maximum display value (all values above are clipped); defaults to 2.5
if \code{slot} is 'scale.data', 6 otherwise}

\item{slot}{Data slot to use, choose from 'raw.data', 'data', or 'scale.data'}

\item{assay}{Assay to pull from}

\item{label}{Label the cell identies above the color bar}

\item{text.size}{text size of gene names}

\item{size}{Size of text above color bar}

\item{hjust}{Horizontal justification of text above color bar}

\item{angle}{Angle of text above color bar}

\item{raster}{If true, plot with geom_raster, else use geom_tile. geom_raster may look blurry on
some viewing applications such as Preview due to how the raster is interpolated. Set this to FALSE
if you are encountering that issue (note that plots may take longer to produce/render).}

\item{draw.lines}{Include white lines to separate the groups}

\item{lines.width}{Integer number to adjust the width of the separating white lines.
Corresponds to the number of "cells" between each group.}

\item{group.bar.height}{Scale the height of the color bar}

\item{legend.remove}{whether remove legend}

\item{legend.title}{legend title}

\item{combine}{Combine plots into a single \code{\link[patchwork]{patchwork}ed}
ggplot object. If \code{FALSE}, return a list of ggplot objects}
}
\value{
A \code{\link[patchwork]{patchwork}ed} ggplot object if
\code{combine = TRUE}; otherwise, a list of ggplot objects
}
\description{
Draws a heatmap of single cell feature expression.
}
