% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{featurePlot}
\alias{featurePlot}
\title{visualize 'features' of single cells in 2D-dimensional space}
\usage{
featurePlot(
  object,
  features,
  colormap = NULL,
  n.colors = 9,
  color.direction = 1,
  show.text.axis = FALSE,
  show.legend = T,
  show.legend.combined = F,
  show.axes = T,
  reduction = NULL,
  group.by = NULL,
  split.by = NULL,
  shape.by = NULL,
  pt.size = 0.1,
  combine = FALSE,
  ...
)
}
\arguments{
\item{object}{seurat object}

\item{features}{Features to plot (gene expression, metrics)}

\item{colormap}{RColorbrewer palette to use (check available palette using RColorBrewer::display.brewer.all()). default will use customed color palette}

\item{n.colors}{Number of colors when setting a color name based on RColorBrewer::brewer.all()}

\item{color.direction}{Sets the order of colours in the scale. If 1, the default, colours are as output by RColorBrewer::brewer.pal(). If -1, the order of colours is reversed.}

\item{show.text.axis}{whther show axis text}

\item{show.legend}{whether show individual legend}

\item{show.legend.combined}{whether just show one legend}

\item{show.axes}{whether show the axes}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{group.by}{Name of one or more metadata columns to group (color) cells by
(for example, orig.ident); pass 'ident' to group by identity class}

\item{split.by}{Name of a metadata column to split plot by;}

\item{shape.by}{If NULL, all points are circles (default).}

\item{pt.size}{size of the dots}

\item{combine}{Combine plots into a single patchworked ggplot object. If FALSE, return a list of ggplot objects}

\item{...}{Extra parameters passed to DimPlot from Seurat package}
}
\value{
ggplot2 object
}
\description{
visualize 'features' of single cells in 2D-dimensional space
}
