% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getEnrichedGO}
\alias{getEnrichedGO}
\title{Perform GO or KEGG enrichment analysis}
\usage{
getEnrichedGO(
  df,
  OrgDb = c("org.Mm.eg.db", "org.Hs.eg.db"),
  category = c("BP", "MF", "KEGG"),
  organism = c("mouse", "hsa"),
  GO.old = NULL,
  idents = NULL,
  invert = FALSE,
  color.heatmap = "Reds",
  n.colors = 8,
  direction = 1,
  dot.size = 3,
  title.name = " enrichment analysis",
  legend.remove = FALSE,
  file.name = "GO_comparison_DEG_clusters",
  width = NA,
  height = NA,
  gg.return = FALSE,
  showCategory = 10,
  pvalueCutoff = 0.01,
  qvalueCutoff = 0.05,
  do.simplify = TRUE,
  simplifyCutoff = 0.7,
  p_val = 0.01,
  text.size = 10
)
}
\arguments{
\item{df}{A data frame containing one column named 'cluster' and one column named 'gene'}

\item{OrgDb}{Genome wide annotation for organism when performing GO analysis}

\item{category}{Type of GO or KEGG enrichment analysis}

\item{organism}{organism when performing KEGG analysis}

\item{GO.old}{Use previously calculated enrichResult object to directly plot the results}

\item{idents}{A vector of cell clusters from df$cluster to perform analysis; default will perform GO comparison analsyis for all the cell clusters}

\item{invert}{invert the cell clusters defined by idents}

\item{color.heatmap}{A character string or vector indicating the colormap option to use. It can be the avaibale color palette in viridis_pal() or brewer.pal()}

\item{n.colors}{number of basic colors to generate from color palette}

\item{direction}{Sets the order of colors in the scale. If 1, the default colors are used. If -1, the order of colors is reversed.}

\item{dot.size}{the dot size}

\item{title.name}{title name}

\item{legend.remove}{whether remove the figure legend}

\item{file.name}{if not NULL, it will automatically save a PDF figure and .csv file of the results}

\item{width, height}{set the figure size when saving figures}

\item{gg.return}{whether return the ggplot object}

\item{showCategory}{Number of GO terms to keep}

\item{pvalueCutoff}{pvalue cutoff on enrichment tests to report}

\item{qvalueCutoff}{value cutoff on enrichment tests to report as significant.}

\item{do.simplify}{whether simply the GO terms by removing redundant terms.}

\item{simplifyCutoff}{cutoff to use for simplifying the GO terms}

\item{p_val}{cutoff associated with each gene when performing the DEG analysis}

\item{text.size}{Font size of the figure}
}
\description{
Perform GO or KEGG enrichment analysis
}
