% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{identifyDEG}
\alias{identifyDEG}
\title{Identify differential expressed genes}
\usage{
identifyDEG(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  group.by = NULL,
  grouping.var = NULL,
  only.pos = TRUE,
  min.pct = 0.25,
  logfc.threshold = 0.25,
  test.use = "wilcox",
  do.plot = TRUE,
  gg.return = FALSE,
  file.name = "heatmap_DEG",
  width = NA,
  height = NA,
  n.top = 10,
  additional.group.by = NULL,
  colors.use = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{ident.1}{Identity class to define markers for. If NULL (default) - use FindAllMarkers.}

\item{ident.2}{A second identity class for comparison. If NULL (default) - use all other cells for comparison.}

\item{group.by}{A vector of variables to group cells by; pass NULL to group by cell identity classes}

\item{grouping.var}{grouping variable when finding the markers that are conserved between the groups, e.g, grouping.var = "batch"}

\item{only.pos}{Only return positive markers}

\item{min.pct}{Threshold of the percent of cells enriched in one cluster}

\item{logfc.threshold}{Threshold of Log Fold Change}

\item{test.use}{which test to use}

\item{do.plot}{whether plot the heatmap}

\item{gg.return}{whether return the ggplot object}

\item{file.name}{if not NULL, it will automatically save a PDF figure of the heatmap}

\item{width, height}{set the figure size when saving figures}

\item{n.top}{the number of genes per cell group shown in the heatmap}

\item{additional.group.by}{A vector of variables to group cells by}

\item{colors.use}{A list of colors to use for the color bar. e.g., colors.use = list(ident = c("red","blue"))}

\item{...}{other parameters in FindAllMarkers, FindMarkers, FindConservedMarkers or doHeatmap}
}
\description{
Identify differential expressed genes
}
