% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vlnPlot}
\alias{vlnPlot}
\title{Violin plot}
\usage{
vlnPlot(
  object,
  features,
  colors.use = NULL,
  colors.ggplot = FALSE,
  idents = NULL,
  group.by = NULL,
  split.by = NULL,
  fill.by = "ident",
  show.text.x = TRUE,
  angle.x = 45,
  hjust.x = 1,
  show.median = FALSE,
  median.size = 1,
  show.boxplot = FALSE,
  pt.size = 0,
  stat.add = FALSE,
  comparisons = list(c("1", "2")),
  stat.method = "wilcox.test",
  label.format = "p.format",
  label.size = 3,
  label.x = NULL,
  y.adjust = 1,
  combine = FALSE,
  show.legend = FALSE,
  ...
)
}
\arguments{
\item{object}{seurat object}

\item{features}{Features to plot (gene expression, metrics)}

\item{colors.use}{defining the color for each cell group}

\item{colors.ggplot}{whether use ggplot color scheme; default: colors.ggplot = FALSE}

\item{idents}{Which classes to include in the plot (default is all)}

\item{group.by}{Name of one or more metadata columns to group (color) cells by
(for example, orig.ident); pass 'ident' to group by identity class}

\item{split.by}{Name of a metadata column to split plot by;}

\item{fill.by}{Color violins based on either 'feature' or 'ident'}

\item{show.text.x}{whther show x-axis text}

\item{angle.x}{angle for x-axis text rotation}

\item{hjust.x}{adjust x axis text}

\item{show.median}{whether show the median value}

\item{median.size}{the shape size of the median}

\item{show.boxplot}{whether show box plot}

\item{pt.size}{size of the dots}

\item{stat.add}{whether add statistical test}

\item{comparisons}{A list of length-2 vectors. The entries in the vector are either the names of 2 values on the x-axis or the 2 integers that correspond to the index of the groups of interest, to be compared.}

\item{stat.method}{a character string indicating which method to be used for comparing}

\item{label.format}{character string specifying label type. Allowed values include "p.signif" (shows the significance levels), "p.format" (shows the formatted p value).}

\item{label.size}{label size}

\item{label.x}{numeric Coordinates (in data units) to be used for absolute positioning of the label. If too short they will be recycled.}

\item{y.adjust}{adjust the y.max when stat.add is TRUE}

\item{combine}{Combine plots into a single patchworked ggplot object. If FALSE, return a list of ggplot objects}

\item{show.legend}{whether show the legend}

\item{...}{Extra parameters passed to VlnPlot from Seurat package}
}
\value{
ggplot2 object
}
\description{
Violin plot
}
